# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "2.3.0",
    "title": "Task Pro API",
    "description": "Task Pro API provides endpoints for managing projects, tasks, and user assignments with secure authentication and efficient workflows.",
    "contact": {
      "name": "TaskPro",
      "url": "https://www.taskpro.qzz.io"
    }
  },
  "servers": [
    { "url": "http://localhost:9537/api" },
    { "url": "https://api.taskpro.qzz.io" }
  ],
  "paths": {
...
    "/user/settings/label": {
      "patch": {
        "tags": ["User"],
        "operationId": "updateLabelSettings",
        "summary": "Update label settings",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "showLabelsOnTask": {
                    "type": "boolean",
                    "example": true
                  },
                  "labelDisplay": {
                    "type": "string",
                    "enum": ["full", "compact"],
                    "example": "full"
                  },
                  "maxLabelsShown": {
                    "type": "string",
                    "enum": ["one", "two", "three", "all"],
                    "example": "three"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/LabelSettingsResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/board": {
      "get": {
        "tags": ["Board"],
        "operationId": "getAllBoards",
        "summary": "Get all boards",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/AllBoardsResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Board"],
        "operationId": "createBoard",
        "summary": "Create new board",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "example": "My New Project"
                  },
                  "icon": { "$ref": "#/components/schemas/BoardIcon" },
                  "background": {
                    "$ref": "#/components/schemas/BoardBackgroundId"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/BoardResponse" },
          "400": { "$ref": "#/components/responses/BoardBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/board/{boardId}": {
      "get": {
        "tags": ["Board"],
        "operationId": "getBoardById",
        "summary": "Get board by id",
        "parameters": [
          {
            "in": "path",
            "name": "boardId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/BoardByIdResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Board"],
        "operationId": "updateBoard",
        "summary": "Update board by id",
        "parameters": [
          {
            "in": "path",
            "name": "boardId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Updated Project Office"
                  },
                  "icon": { "$ref": "#/components/schemas/BoardIcon" },
                  "background": {
                    "$ref": "#/components/schemas/BoardBackgroundId"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/BoardResponse" },
          "400": { "$ref": "#/components/responses/BoardBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Board"],
        "operationId": "deleteBoard",
        "summary": "Delete board by id",
        "parameters": [
          {
            "in": "path",
            "name": "boardId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "The resource was deleted successfully." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/column/{boardId}": {
      "post": {
        "tags": ["Column"],
        "operationId": "createColumn",
        "summary": "Create new column",
        "parameters": [
          {
            "in": "path",
            "name": "boardId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "example": "To Do"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/ColumnResponse" },
          "400": { "$ref": "#/components/responses/ColumnBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/column/{boardId}/order": {
      "patch": {
        "tags": ["Column"],
        "operationId": "updateColumnsOrder",
        "summary": "Update columns order",
        "parameters": [
          {
            "in": "path",
            "name": "boardId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ids"],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "67470731ed1b2ff3a2e5bd56",
                      "67470773c750b6298e1892ad",
                      "67470737ed1b2ff3a2e5bd57"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/UpdateColumnsOrderResponse"
          },
          "400": { "$ref": "#/components/responses/UpdateOrderBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/column/{columnId}": {
      "patch": {
        "tags": ["Column"],
        "operationId": "updateColumn",
        "summary": "Update column by id",
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Doing"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/ColumnResponse" },
          "400": { "$ref": "#/components/responses/ColumnBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Column"],
        "operationId": "deleteColumn",
        "summary": "Delete column by id",
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "The resource was deleted successfully." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/label": {
      "get": {
        "tags": ["Label"],
        "operationId": "getAllLabels",
        "summary": "Get all labels",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/AllLabelsResponse" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["Label"],
        "operationId": "createLabel",
        "summary": "Create new label",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "color"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "Bug"
                  },
                  "description": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Description of the bug"
                  },
                  "color": { "$ref": "#/components/schemas/AccentColor" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/LabelResponse" },
          "400": { "$ref": "#/components/responses/LabelBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/label/{labelId}": {
      "patch": {
        "tags": ["Label"],
        "operationId": "updateLabel",
        "summary": "Update label by id",
        "parameters": [
          {
            "in": "path",
            "name": "labelId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "Bug"
                  },
                  "description": {
                    "type": ["string", "null"],
                    "minLength": 3,
                    "example": "Description of the bug"
                  },
                  "color": { "$ref": "#/components/schemas/AccentColor" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/LabelResponse" },
          "400": { "$ref": "#/components/responses/LabelBadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Label"],
        "operationId": "deleteLabel",
        "summary": "Delete label by id",
        "parameters": [
          {
            "in": "path",
            "name": "labelId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "The resource was deleted successfully." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Board": {
        "type": "object",
        "required": [
          "id",
          "title",
          "icon",
          "background",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672f1d2f02c8b04130c718c"
          },
          "title": {
            "type": "string",
            "example": "Project office"
          },
          "icon": { "$ref": "#/components/schemas/BoardIcon" },
          "background": { "$ref": "#/components/schemas/BoardBackground" },
          "userId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          },
          "columns": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Column" }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          }
        }
      },
      "BoardBackground": {
        "type": "object",
        "required": ["identifier", "url"],
        "properties": {
          "identifier": { "$ref": "#/components/schemas/BoardBackgroundId" },
          "url": {
            "type": ["string", "null"],
            "format": "url",
            "example": "https://res.cloudinary.com/dmbnnewoy/image/upload/v1707099093/TaskPro/board_bg_images/desk/nfxep55xgvpq7xitemq1.jpg"
          }
        }
      },
      "BoardBackgroundId": {
        "type": "string",
        "enum": [
          "default",
          "flowers",
          "mountains-night",
          "sakura",
          "night-sky",
          "tropics",
          "light-sky",
          "sea-rocks",
          "blue-ball",
          "red-moon",
          "sea-boat",
          "air-balloons",
          "gorge",
          "sea-boat-2",
          "air-balloons-2",
          "northern-lights"
        ],
        "example": "flowers"
      },
      "BoardIcon": {
        "type": "string",
        "enum": [
          "sparkles",
          "layout",
          "kanban",
          "layers",
          "compass",
          "target",
          "sunrise",
          "orbit",
          "palette",
          "leaf"
        ],
        "example": "grid"
      },
      "Column": {
        "type": "object",
        "required": [
          "id",
          "title",
          "order",
          "boardId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672f604c07147fc7ae1bb81"
          },
          "title": { "type": "string", "example": "In progress" },
          "order": { "type": "integer", "example": 1 },
          "boardId": {
            "type": "string",
            "example": "67470718ed1b2ff3a2e5bd55"
          },
          "tasks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Task" }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          }
        }
      },
      "Task": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "priority",
          "deadline",
          "completed",
          "columnId",
          "order",
          "completedAt",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "title": {
            "type": "string",
            "example": "Publication of the project"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Review the project materials: Familiarize yourself with the project's content, including text, images, and any supplementary materials."
          },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "deadline": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": "2025-10-15T00:00:00.000Z"
          },
          "order": { "type": "integer", "example": 1 },
          "completed": { "type": "boolean", "example": false },
          "completedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": null
          },
          "columnId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          },
          "labels": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Label" }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          }
        }
      },
      "TaskPriority": {
        "type": "string",
        "enum": ["without", "low", "medium", "high"],
        "example": "high"
      },
      "Label": {
        "type": "object",
        "required": ["id", "name", "color", "userId", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "name": {
            "type": "string",
            "example": "Bug"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Description of the bug"
          },
          "color": { "$ref": "#/components/schemas/AccentColor" },
          "userId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          },
          "taskId": {
            "type": ["string", "null"],
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-29T14:30:00.000Z"
          }
        }
      },
      "AccentColor": {
        "type": "string",
        "enum": [
          "blue",
          "purple",
          "red",
          "gray",
          "cyan",
          "yellow",
          "indigo",
          "green"
        ],
        "example": "red"
      },
...
      "LabelSettings": {
        "type": "object",
        "required": [
          "id",
          "showLabelsOnTask",
          "labelDisplay",
          "maxLabelsShown",
          "userId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "6672fdccc07147fc7ae1bb93"
          },
          "showLabelsOnTask": {
            "type": "boolean",
            "example": true
          },
          "labelDisplay": {
            "type": "string",
            "enum": ["full", "compact"],
            "example": "full"
          },
          "maxLabelsShown": {
            "type": "string",
            "enum": ["one", "two", "three", "all"],
            "example": "three"
          },
          "userId": {
            "type": "string",
            "example": "6671743db85c96453e04a684"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer", "example": 400 },
          "message": {
            "oneOf": [{ "type": "string" }, { "type": "object" }]
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": { "statusCode": 404, "message": "Not found" }
            }
          }
        }
      },
...
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": { "statusCode": 401, "message": "Unauthorized" }
            }
          }
        }
      },
...
      "UpdateOrderBadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": {
                "statusCode": 400,
                "message": {
                  "ids": "The field must be of type array"
                }
              }
            }
          }
        }
      },
...
      "LabelSettingsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "$ref": "#/components/schemas/LabelSettings"
            }
          }
        }
      },
...
      "AllBoardsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/Board" }
            }
          }
        }
      },
      "BoardResponse": {
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Board" }
          }
        }
      },
      "BoardBadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": {
                "statusCode": 400,
                "message": {
                  "title": "The field must be at least 3 characters",
                  "icon": "The field must be one of the following values: sparkles, layout, kanban, layers, compass, target, sunrise, orbit, palette, leaf"
                }
              }
            }
          }
        }
      },
      "BoardByIdResponse": {
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                { "$ref": "#/components/schemas/Board" },
                {
                  "type": "object",
                  "properties": {
                    "columns": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          { "$ref": "#/components/schemas/Column" },
                          {
                            "type": "object",
                            "properties": {
                              "tasks": {
                                "type": "array",
                                "items": { "$ref": "#/components/schemas/Task" }
                              }
                            }
                          }
                        ]
                      }
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "ColumnResponse": {
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Column" }
          }
        }
      },
      "ColumnBadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": {
                "statusCode": 400,
                "message": {
                  "title": "The field must be at least 3 characters",
                  "boardId": "The field must have a valid ObjectId format"
                }
              }
            }
          }
        }
      },
      "UpdateColumnsOrderResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/Column" }
            }
          }
        }
      },
...
      "AllLabelsResponse": {
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/Label" }
            }
          }
        }
      },
      "LabelResponse": {
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Label" }
          }
        }
      },
      "LabelBadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse",
              "example": {
                "statusCode": 400,
                "message": {
                  "color": "The field must be one of the following values: blue, purple, red, gray, cyan, yellow, indigo, green"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}